Move CAS2 OASys Section Transform Logic
This moves the logic into a CAS2-speciifc class, as it’s implementation differs from other CASs
